requirements.txt
opencv-contrib-python
PyOpenGL
PyOpenGL_accelerate
pyrr
glfw
tqdm
matplotlib
streamlit
